Various new methods in RNG, better docs
There was some incorrect information in some docs for RNG, I fixed it
where I found it.  There's a shuffleInPlace() method for List now, not
just arrays. There's nextFloat() that takes an outer bound. There's
minIntOf, minLongOf, minFloatOf, and minDoubleOf methods, plus max
versions as well, which produce multiple random numbers between 0 and a
bound and return the min or max of that bunch of numbers.