Update to libGDX 1.9.8, improve NumberTools
The change to libGDX 1.9.8 is generally minor and no issues have
surfaced so far, but because libGDX's ScrollPane.setWidget() is now
deprecated and it may be removed in an upcoming release, the
differently-named and just-added ScrollPane.setActor() is used instead.
This setActor() method wasn't available in libGDX 1.9.6, so as of this
revision the display module needs libGDX 1.9.7 or newer to build.
Because 1.9.8 seems pretty painless to upgrade to, I think this won't be
an issue. It's hard to measure, but it looks like the framerate is
smoother in this libGDX version; there may have been optimizations that
reduce GC load.

The NumberTools changes affect the randomFloat, randomDouble,
randomSignedFloat, and randomFloatCurved methods. These simply use a
more efficient random number generation algorithm, but operate on long
values instead of the ints they previously used so the issues with
larger-than-normal ints on GWT won't change results. They also may
generate the full range of numbers from 0.0 to 1.0 instead of only a
fraction of it, which they were limited to before.