Fix the FOVCache benchmark's dead code elim
Dead code elimination was probably removing most of the code that the
benchmark was supposed to test, so now this is all a little more robust.
If you actually run the test, you'll notice a hefty delay before the
measureCachedFOV benchmark starts; this delay isn't recorded in the
benchmark results, but it should be considered if evaluating FOVCache.
To make it easier to decide, FOVCache is probably going to be removed
because it's so much slower than the alternatives, adds a big chunk of
startup time, and isn't GWT-compatible.