Add ORK dialect thing in Thesaurus. Various stuff.
Using Thesaurus.ORK.process() on the first sentence of The Wizard of Oz
yields: "Doroddy lived in da midst of da great Kansas prairiez, widd
Uncle Henry, who was a farmah, an' Aunt Em, who was da farmah's wife."
This is meant to evoke a specific group of Orks with a 'k' from a
specific game, but it's up to you what it could apply to. Please don't
make it a stereotype of a real-world (especially contemporary) group,
though.

OrderedMap.putPairs() should be useful.

FakeLanguageGen.SIMPLIFY_ARABIC, which is still deprecated, mistakenly
was using a default value equal to FakeLanguageGen.LISP. Now the default
Modifier does nothing, so SIMPLIFY_ARABIC does too (it shouldn't ever be
needed anyway).