Remove several (bad) RandomnessSources
Bard, Beard, Bird, Herd, Horde, Slap, and Zap RNGs were removed, because
they were all tested with far-too-lenient standards and in general are
slower than they should be with such dubious qualty. XkcdRNG was moved
to squidlib-performance because it really is just a gag RNG, but can be
useful in benchmarking, so it's only available during benchmarking now.

JabRNG is new; it has 63 bits of state (the LSB is always 1, so the
state is always an odd number) and is extremely fast but doesn't yet
fully pass PractRand. VortexRNG has also been added; it has two 64-bit
longs for state and updates them both by different amounts, for a period
of 2 to the 64 but also 2 to the 64 possible streams.