Add Noise.Basic1D because we need a 1D noise impl
I don't think it really needs its own file, because it's a tiny piece of
code. SquidPanel has a small change to the defaults that shouldn't
affect anything visually (all characters default to transparent, but
they were space chars, which still involve BitmapFont to draw, instead
of the NUL char '\u0000' or '\0', which renders as a block like
backgrounds usually use for SquidPanel but could be slightly faster
because it doesn't need a GlyphLayout or anything like it). If you
already fill a SquidPanel up with chars, then the default chars in the
contents array won't matter to you anyway.