Adjust IColoredString.wrap(). Resolves #194 .
TinyTest shows the fix works, where it didn't before. The issue has to
do with String.split()'s behavior when the only contents are a trailing
delimiter, vs. StringKit.split()'s behavior. We now use
StringKit.split() because before, String.split() could "eat" the space
used as a delimiter if it was the full contents of an
IColoredString.Impl bucket. A check for this was present and worked in
any case where there was other text along with the space, but it
wouldn't execute if the full text was a single space.