Improve layered noise, WorldMapGenerator options
Noise.Layered(n)D now can optionally set the lacunarity of the noise,
which is how much the frequency should change with each octave. The API
stays the same for existing code.

WorldMapGenerator now has an experimental mode,
WorldMapGenerator.SphereMapAlt, that uses something like the distortion
in SphereMap (applying a Noise.Ridged3D as a modifier to the x-position
of a Noise.InverseLayered3D call), but instead of modifying x, it takes
the result of a call to a Ridged3D with x, y, z as inputs, and uses the
same x, y, and z as inputs along with that Ridged3D result in a call to
Noise.Layered4D. This seems to be a decent way to reduce the frequency
and severity of ring artifacts, where a band of high or low terrain
shows up in a non-natural-seeming way. There's an example at
https://tommyettinger.github.io/DorpBorx/worlds6/ .