Slight improvement on trig-heavy code speed
Calling sin() and cos() isn't a major bottleneck for much of SquidLib,
but seamless noise and a few other places make a lot of trigonometric
method calls. We now use a very fast approximation to sin() in those
places, because we don't need a perfect circle when almost everything is
grid-shaped or world-like where this is used, and the approximation is
more than close enough to pass for a circle in a roguelike. Each
approximation in NumberUtils is also about 40x faster than Math.sin() or
Math.cos() in benchmarking (in a related project, tommyettinger/Sarong
), and this makes a small difference in DetailedWorldMapDemo. The
difference may be insignificant when using a lot of octaves of noise,
because those take much more time to calculate, but probably is more
significant on smaller maps with lower detail levels.

This builds the changed docs as well, which turns out to be a lot of
files.