NumberTools.swayRandomized() works!
This conceptually-simple form of 1D continuous noise turrned out rather
hard to implement, but it runs now and a live sample is shown in
HashVisualizer (no keys currently need to be pressed). It smoothly
transitions between random peaks and valleys, sometimes staying close to
level, and takes a seed for customization. It should be fairly fast,
though there are ways to speed it up in the future if more is known
about the parameters it will get.

WhirlingNoise now avoids a step at startup by using already-calculated
constants instead of calculating them at runtime. MerlinNoise respects
the state given to it now, and there's an example showing river-like
shapes being formed from the boundaries of MerlinNoise calls.