Make WeightedTable a little "lighter-weight"
It doesn't actually need to internally store a double array, and it was
converting 32 random bits to a double for comparison with a value that
didn't really need to be a double. So now it stores one int array that
interleaves probability (as a positive int) with alias (referring to an
index in the original data given to the WeightedTable), and the
resulting code can potentially be easier to make a lot of (half as many
arrays per WeightedTable, and potentially the only other field (size)
could be stripped out to make the class into static methods that get
random values using only an int array, if you had a reason to manipulate
thousands of tables). This does change what values WeightedTable
returns, but not how much it changes state (potentially limiting the
effects this change has on a game's pseudo-random values from other
generators).