Add methods to get "distance metrics" for colors
These are used in Filters.PaletteFilter to more accurately adjust colors
to match a palette. MapUtility also now will use a uniform lightness if
no frame is given, which is probably more suitable for limited-palette
filters (before, it would act like it was using PerlinNoise to distort
the lightness over time, but wouldd be frozen in one frame, leading to
an odd appearance when deep water sometimes used one palette color and
sometimes used another in the same area).