Fix seeding issue with swayRandomized, add language
FakeLanguageGen.INSECT is meant for the frequently-seen fantasy or alien
species with insectoid characteristics, and should sound fairly unique.
NumberTools.swayRandomized() wasn't taking the seed into account
correctly, effectively treating a change of +1 to seed as +1 to the
floor value (which would be encountered soon after some progress is made
along the 1D line), instead of an actual change to the generated
sequence. HashVisualizer is getting some tinkering as well for new types
of noise.