Add sin/cos in degrees to NumberTools, speedup
But the speedup isn't enough to beat MathUtils.sin() or .cos() from
libGDX, with its medium-large lookup table. Currently, MathUtils.cos()
is about 3x faster than NumberTools.cos(), and the speed is almost
identical for the version in degrees. A recommendation has been added to
consider MathUtils if possible in user code; however, it is possible the
lookup table MathUtils uses might not fit in a processor cache on a busy
machine, and in that case NumberTools may do better.