Update WhirlingNoise point hash technique
This gives a very nice speedup on 3D and 4D noise, making WhirlingNoise
the fastest 3D continuous noise we have (and in a tie with SeededNoise
for the fastest 4D noise, with WhirlingNoise at an advantage). This
still uses the natural-looking gradient vectors that make WhirlingNoise
so good for world maps. New previews are up in
http://squidpony.github.io/SquidLib/SphereWorld3.png and
http://squidpony.github.io/SquidLib/TilingWorld3.png .