Working (flawed) perceptual hash for GreasedRegions
It currently doesn't seem very precise, hashing in lumps of 4 bits at a
time, all on or off. It's ported from
https://github.com/commonsmachinery/blockhash , which seems good but
works on RGBA images, so some tweaks may be needed.
DetailedWorldMapWriter uses the perceptual hash to compare generated
maps with Earth, and so far it's found one that it thinks is very
similar... but isn't really in practice. Maybe it could be adapted,
though.