LineKit can remove unseen branches on lines
The LineKit.pruneLines() method is a little complex, but it should be
handy for games that use box-drawing characters for maps. You can see
how it's used in SparseDemo; the requirement to use a GreasedRegion to
store seen tiles (needed by LineKit.pruneLines() ) may be relaxed to
allow other types in the future.