Add ThrustAlt32RNG, change some hashing internals
ThrustAlt32RNG is optimized for GWT and, unlike previous 32-bit RNGs,
care has been taken to account for the utter strangeness of the int type
on GWT. In this new RandomnessSource, multiplication is always by at
most 20 bit ints when producing an int (which keeps 32-bit inputs within
the 53 bits at which numbers are precise on GWT). In the super-sourced
version of ThrustAlt32RNG, multiplications, additions, and other
potential overflow sources are followed by `| 0`, which is a no-op on
desktop but ensures the int is actually in the 32-bit range on GWT.

HashCommon and some IHashers in CrossHash now also refrain from
multiplying by a large int when mixing bits; a new (slightly slower)
version of the mixing code is used that multiplies by a 19-bit number
instead of the earlier 32-bit one, then XORs that result with two
different bitwise rotations of itself. Desktop JVMs tend to be very good
at optimizing bitwise rotations (into SSE intrinsics, as far as I
recall), and since this will now behave correctly on GWT, OrderedMap and
other classes that depend on this mixing should have their hash tables
structured as they expect them, rather than in some glitchy alternative.

PintRNG isn't used any more in two places, TableSet and TuringPattern,
because it doesn't behave the same on GWT as on desktop or Android; the
usage there has been replaced with ThrustAlt32RNG.