Zooming works better in MimicMap
This used similar code to the one used to zoom the GreasedRegion in
FantasyPoliticalMapper, and seems to work except at very high zoom
levels (other map types have a similar issue). There's also a
diamond-shape bias in land shapes on zoom, which is kinda unavoidable
for now due to how they are zoomed in (we can't just add more map data
to mimic; we're near the limit for a UTF-8 String already in the source
for WorldMapGenerator where we have a compressed-format map of Earth).