Change Noise.Basic1D to a better-looking algo
It may also be faster in some usages. The output of 5 octaves layered
looks like https://i.imgur.com/7SW2Lol.gifv , though this uses
getNoise() instead of getNoiseWithSeed() because the latter can be
slower with this method. The idea is to generate 4 random doubles and
use them as frequencies when calling an ordinary sway() method
(specifically, a cubic-interpolating variant). If the seed doesn't
change, the doubles don't need to be regenerated, and this may lead to
nice speed boosts.