Additions to Mnemonic, Thesaurus
You can now generate a considerably more-memorable mnemonic, such as
"Ensorcelled assassin and the king's rifle", from an int, in this case
-1269126816 . The existing Mnemonic API is still there, and the
mnemonics using words depend on Thesaurus, so if or when Thesaurus gets
more adjectives and/or nouns (as in this commit), the mnemonics will
change. A constructor can probably be provided soon that will use only a
specified list of adjectives and nouns.