Restructure ObText a little; add useful methods
You can now take an ObText instance and produce an ArrayList of String,
a Set of String (UnorderedSet or OrderedSet), a Map of String to String
(UnorderedMap or OrderedMap), or a Map of String to ArrayList of String
(UnorderedMap or OrderedMap).  ObText now inherits from ArrayList
instead of AbstractList (where it just used an ArrayList internally
anyway), and this may enable some future cleanup. ObText.ObTextEntry now
specifies its associated entries as an ArrayList instead of a List, but
it may be able to specify ObText instead to better support some
recursive things.