Add Zag32RNG (high-quality and very GWT-friendly)
Zag32RNG is super-sourced with a minor set of changes for GWT, and uses
no operations that will misbehave on GWT relative to desktop. It can be
over 30x times as fast as long-based generators like LightRNG on GWT
when generating int values and is often about 2x to 3x as fast when
generating long values, and should also perform well on 32-bit mobile
devices or older PCs.

This commit also adjusts the return types of copy() methods on
RandomnessSource implementations so they return their actual type and
not RandomnessSource; this should make copying things like LightRNG to
get another LightRNG cleaner and not require casting. I think some
classes had this done before, but not all.