Add Iosevka Slab family (MSDF), change MSDF shader
This commit also adds a method to IColoredString so you can wrap an
IColoredString into an existing List of such IColoredString, using the
technique with a `buf` parameter I've seen used by @smelc and that I
think he will appreciate seeing in his class. As for the MSDF stuff,
other than the new font, the existing MSDF family font had its baseline
adjusted, and the shader tweaks seem to help downscaled MSDF renders for
some amount of downscaling, while certainly making curves smooth on
upscaled renders. The weight may need to be tweaked if the new stuff
appears too bold or too light. SparseDemo has been updated to use a
TextFamily and show italic text for Sun Tsu's name as well as
translations of it.