Add fast, not-so-precise atan2 approximation
Since almost everything in SquidLib is on a grid, extreme precision
isn't especially necessary for atan2(), so we can get by with a
relatively imprecise method in NumberTools that is 10-20x faster than
Math.atan2(). This is also faster than MathUtils' approximation in
libGDX, by about 3x, but still has a worse maximum error and probably
also average error. The maximum error is under 0.07 radians for the
result, which is unlikely to have much noticeable effect on a discrete
grid.