Add Oriole32RNG for GWT randomness
Oriole32RNG has a fairly large period of 0xFFFFFFFFFFFFFFFF00000000 (but
too large of a state to be a StatefulRandomness), passes 32TB of quality
tests, and is rather fast on GWT, while still being a little faster than
Zag32RNG on desktop. It's almost as fast as ThrustAlt32RNG on GWT, but
has a much more useful-in-practice period and can produce all int
values, which is a severe issue for ThrustAlt32RNG for some usage. If
you don't need a StatefulRandomness or SkippingRandomness (or
determine() methods), Oriole32RNG may be the best choice for
GWT-targeting applications.