Generalize APIs from RNG to IRNG (RNG still works)
Not all could be generalized and there are still some doc warnings from
broken links, but this should allow more unusual or optimized IRNG types
to be usable alongside RNG.

RNG.nextInt(int) has changed its implementation to use
RandomnessSource.next() instead of nextLong() and to always make one
call to it (no branching, and no modulus either). This probably will
change results of calls to RNG.nextInt(int), but it makes code that
previously used RNG.nextIntHasty() no longer need to much of the time
(nextIntHasty still handles negative bounds differently). This is
important because IRNG doesn't have a nextIntHasty method.

This may break some user code if it previously obtained an RNG object
from a method like getRandom() on a dungeon generator where that dungeon
generator now uses and returns an IRNG from getRandom(), not an RNG.
There aren't many places where this happened, and none of them are
probably very hard to work around... Feel free to post a GitHub issue if
this commit is an inconvenience or breaks code in a hard-to-fix way. The
biggest issue I see is that IRNG isn't Serializable, and squidlib-extra
can't serialize an unknown implementation of an interface. smelc's
solution of toSerializable() may be the best way.