Make IRNG a RandomnessSource, move obscure methods
RNG.randomInterleave() and RNG.approximateBits(int) have been moved to
be static methods in GreasedRegion (the only place I know of where they
are used) and changed to take any RandomnessSource, instead of just an
RNG. RNG is now also a RandomnessSource, since it has all the right
methods. The versions of approximateBits() and randomInterleave() in RNG
are deprecated but not removed yet.