Major improvement in SquidInput: Key repeat
It took me a long time to realize this was an issue, since I almost
always use the mouse, but precise key repeating really does improve the
experience on a keyboard. I still need to check how this works on GWT,
and the implementation may become more configurable. Also, SquidInput is
affected, but I haven't changed SquidKey or VisualInput, though the
latter may have inherited changes.