Elias (LOS) is now GWT-compatible, faster
Elias (the class in squidmath) and LOS with ELIAS mode (in squidgrid)
now no longer use threads and aren't synchronized, which, curiously,
leads to a significant speedup. With an updated LOSBenchmark, we can see
that Elias is, compared to the current fastest LOS type DDA, about 3x
slower instead of double-digits slower like it was before. It should
produce a high-quality line that can peek around corners symmetrically,
so the smallish slowdown may be worth it if the quality of lines matters
more than how fast you can get them.