Move some data structures to squidlib-extra
BundleBiMap, BundleBundleBiMap, MultiArrangement, and MultiKey are now
in the same package but require a dependency on squidlib-extra instead
of always being in squidlib-util. This also deletes TableSet, which was
unused and never finished, and Isaac32RNG, which was not suitable on GWT
and that was the only place it would have been relevant. SquidID has
changed algorithm to use the unique states of xoroshiro128+, so there
are guaranteed to be `pow(2, 128) - 1` IDs it can generate before
repeating any of them. Since nobody has even `pow(2, 64)` bytes of
memory available yet, there's no way that many IDs will be exhausted for
a long time.

MetsaMapFactory is being re-examined to see if it can be used or if it
is superceded by WorldMapGenerator.  The demo for that has been moved to
MetsaWorldMapDemo, in squidlib's tests directory.