Tough tweaks to WMG Elliptical/SpaceView
It's hard to tell if these are improving the generated maps, but it
seems to me like the changes do help with making believable maps. You
can compare the old elliptical maps at
https://tommyettinger.github.io/DorpBorx/worlds11/index.html with the
new elliptical maps at
https://tommyettinger.github.io/DorpBorx/worlds12/index.html ; the goal
of the changes in 12 is mainly to reduce the frequency of chaotic areas
where height spikes very high and very low in a small region, but also
to reduce the number of mid-sized non-continent islands and lakes. Both
of these seem to be better in the newer version. There's a new
wallpaper-sized preview in
https://squidpony.github.io/SquidLib/EllipseWorldMap.png as well, with
the improved color scheme and the current ellipse algorithm.