Major quality improvement on NumberTools.atan2()
For 65536 points, the cumulative error for the old NumberTools.atan2()
was about 2900, while it is about 4 for the new one. It's about half the
speed, on-par with libGDX's approximation but significantly better on
error (cumulative error for GDX's MathUtils.atan2() is about 170 in this
case). Notably, WorldMapGenerator.SpaceViewMap can switch to
NumberTools.atan2() now without incurring visual flaws, though it is
unclear if a noticeable performance increase comes with it.