More DijkstraMap optional optimizations
You can now optionally pass in an ArrayList of Coord for the various
findPath() family of methods to append to, instead of making new
ArrayLists every time a path was requested. There are some other
optimizations in here, but I may have missed some behavior changes in
some of the DijkstraMap methods. These should be minor at worst, but if
there's an issue, there's a section on SquidLib's repo just for that.