Add LinnormRNG. It's really good.
It passes PractRand at 32TB with no anomalies (which no other generators
have done that I have tested). It is faster than either LightRNG or
XoRoRNG (but slightly slower than ThrustAltRNG). Unlike ThrustAltRNG, it
can produce all possible long values, instead of only some. It is a
StatefulRandomness. In most cases, it seems like it is a better default
than LightRNG, unless the SkippingRandomness features are needed (and we
have MoonwalkRNG if you need that and more)... it may be made the
default for RNG and StatefulRNG if only for the nice speed boost, since
it's otherwise identical to LightRNG in terms of space usage, and
quality seems (after about 2 days of testing) very good. If not, there
may be a new class that extends AbstractRNG and is hardcoded to use
Linnorm with the extra features mentioned in its JavaDocs (like
MoonwalkRNG does with LightRNG).

Also, dragons.