Experiment with crisp outline shaders,  QRNGs
The new shader in DefaultResources can be used as shown in ShaderDemo,
and looks a little better than the earlier one (in part because it is
always used with a bold font). VanDerCorputQRNG.altDetermine() has its
algorithm moved to planarDetermine(), because the old way had a strong
planar bias (putting all points on parallel lines in 2D). The new way is
more random but also less well-spread-out. LinnormRNG is trying out the
technique for generating a long between 0 and a target value from
libGDX's RandomXS128.