Improve sharpness on DistanceFieldTest, TCF stuff
TextCellFactory now has a resetSize() method that can be used to adjust
the size of a TextCellFactory every frame if needed; before, calling
initBySize() too often could yield a driver crash for my Intel graphics
that would cause a Java minidump. This was discovered in
DistanceFieldTest, which now resizes the font as you probably normally
would if scaling some glyphs but not others (by setting the size on the
font), instead of by changing the viewport as before. This leads to much
more clear and crisp fonts in both MSDF and regular SDF modes.