Add NotoSerif font, help IColoredString render
NotoSerif is a variable-width MSDF font meant to be used in TextPanel;
it looks pretty nice. NotoSerif includes some workarounds for odd
behavior in BitmapFont, some of which may have been related to
IColoredString,presentWithMarkup() putting in an extra `[]` that wasn't
needed and caused BitmapFont to act strangely. GreasedZOI may need some
work; it was never really tested out fully when it was introduced.