SColor stores its representation as a packed float
There's a lot of places where SColor.toFloatBits() gets called and it
was doing the same work every time; now we store this value in SColor,
adding some extra overhead to each SColor (one float) but hopefully
speeding up some code that is called many times each frame. This float
is recalculated when SColor.clamp() is called, which might not be as
often as I hope it is, but changing SColors by mutation is generally
rather rare. You can call clamp() manually to ensure the packed float
and other RGBA channel floats are in-range.