ColorCenters were allocating TONS of Longs
The many Long objects were getting GCed quickly enough, but it was probably slowing down Android a lot when SquidColorCenter was in use (or any IColorCenter.Skeleton). Now a very weird custom hashing situation is used, using UnorderedSet with a custom IHasher that uses the IColorCenter's definition of equality to figure out how things are hashed. This isn't the best example of a custom hasher, but it's definitely a case where a HashMap of Long to T was causing trouble and an UnorderedSet of T with custom hashing doesn't seem to.