Change how land is calculated in WMG
If you change waterModifier (which was poorly named, and is now called landModifier) to higher than 1.0 when you call WorldMapGenerator.generate(), you can now get maps that are all land, like Mars or most mid-sized exoplanets. You can also set it close to 0 to get almost-all-water or all-water maps. DetailedWorldMapDemo currently is testing a "anime-style" flat color effect by seriously reducing the possible colors (instead of 256 values for red, green, and blue individually, there are currently 6 values possible for red and for blue, and 7 for green).