Fix random bounded nextLong, RegExodus update
Due to wishful thinking on my part, the nextLong(long) methods on various RandomnessSources, RNG, and AbstractRNG all failed to produce random longs very close to the bound, so small bounds would never see some results they would expect. This seems to be fixed, and I went through the tests a lot more thoroughly this time. RegExodus is updated to 0.1.10, which is on Maven Central, and this will be needed if/when libGDX starts using GWT 2.8.1 or 2.8.2 (that release of RegExodus is a compatibility release for GWT 2.8.2, and also includes fixes).

I encourage users to update their projects to use RegExodus 0.1.10 instead of 0.1.9, which in SquidSetup projects is a tiny change to regExodusVersion in
gradle.properties in the project root folder.