Use String.hashCode() in Falcon
This is somewhat experimental on GWT until I can confirm that String.hashCode() acts according to its contract on desktop JDKs. If it works, it will be much faster than hashing each char in a String normally. If it doesn't work, Falcon isn't currently used in SquidLib.

This also updates HashBenchmark.