Allow rotating EllipticalMap
Also some improvements to both EllipticalMap and SpaceViewMap, which now use about the same algorithm and now might produce the same world given an identical seed and parameters. This algorithm is a little faster because it can avoid generating 4D noise. Some of the trigonometry calls are also faster, taking advantage of how sin() and cos() allow large inputs and don't need it wrapped with modulus.