Big experiment with GreasedRegion
I'm trying to avoid unnecessary allocations in getter-type methods, since GreasedRegion.singleRandom() on a 400x400 area was profiling as being about 100x slower than pathfinding with DijkstraMap on a 400x400 map. This whole large change is mostly as simple as only calculating the total size and sizes of sections when needed, but was much more involved in some locations. This storing of size info (rather than remaking and discarding often) makes each GreasedRegion about 1.5x larger, but we were never using more than a relative few at a time anyway.