Fix build, spill() behavior
MaskedShapeGenerator, which is dubiously useful currently, was using protected members of GreasedRegion without really needing to; it uses an alternative to the protected field. This fixes the compilation errors on JitPack.

The tests now all pass again, and the tests that allow printing to be configured all look into one place, LanguageGenTest.PRINTING, to see if they should spew text into the terminal while they test. DiceTest wasn't working because I needed to specify LightRNG as the RandomnessSource to maintain LightRNG results. GreasedRegionTest wasn't working because I tried to get clever with spill()'s implementation, and it needed serious debugging to track down a cause (there were several, with varying occurence frequencies, but spill() is pretty much the only method I tried to do anything fancy on in the last commit).