Replace collection first and last fields with order
This applies to OrderedMap, OrderedSet, and Arrangement. We already track ordering in the order IntVLA, so tracking the first and last items in the ordering separately is error-prone and has caused bugs with alter(). Now code in those three data structures uses the first item in order instead of first, or the last item in order instead of last. This saves some frequent checks for "is this the first or last item, if so set first and/or last", and is conceptually cleaner. It also seems to fix a bug encountered with OrderedMap in my ColorGuard project.