Add NumberTools.atan2Rough
After benchmarking various atan2() approximations in the Sarong project's benchmarks (which are easier to manage than SquidLib's), I found out that SquidLib's approximation in NumberTools.atan2() is unusually precise (the best of any approximation that was significantly faster than Math.atan2()), but is slower in doing so. It isn't the slowest one benchmarked, but it's one of the slower ones. So I looked around and found a second-order, as opposed to seventh-order, approximation, and now SquidLib provides that as an option called atan2Rough().