Bench point hashes, SColor stuff
The color "TEA_GARDEN_CONTEMPLATION" was incorrectly named "Tead Garden Contemplation",  so now the old one is moved to the compatibility section and the correct color is called "Tea Garden Contemplation". Surprisingly, Noise.PointHash is a little faster than Noise.HastyPointHash, and the speedup is consistent across dimensions. This is unusual because PointHash is doing only 64-bit math while HastyPointHash is doing some 32-bit math, and less of it (but HastyPointHash needs array lookups). Full benchmarks on point hashing will be run once an RNG test finishes tomorrow, since it's putting a lot of load on this computer currently.